Broke out findChargePath and dependent isValidChargePath, refactoring the former to pass in the board. #28
