+++ client/utils/find-charge-path.js
@@ -1,0 +1,37 @@
+const getValidChannels = require('./get-valid-channels.js')
+const findAdjacentTiles = require('./find-adjacent-tiles.js')
+const getOppositeDirection = require('./get-opposite-direction.js')
+
+function findChargePath(board, chargeCoordinates) {
+  for (let i = 0; i < board.length; i++) {
+    for (let j = 0; j < board[i].length; j++) {
+      board[i][j].chargeStatus.chargeAligned = false
+    }
+  }
+  let inChargePath = chargeCoordinates
+  let lastChargedTile = 0
+  while (inChargePath) {
+    let adjacent = findAdjacentTiles(board, inChargePath)
+    let validChannels = getValidChannels(board, inChargePath)
+    lastChargedTile = inChargePath
+    for (let i = 0; i < validChannels.length; i++) {
+      let channelOpposite = getOppositeDirection(validChannels[i])
+      if (adjacent[validChannels[i]]) {   // if a tile exists in the direction of a valid channel
+        let adjacentTile = board[adjacent[validChannels[i]][0]][adjacent[validChannels[i]][1]] // adjacentTile is assigned the value of tile
+        if (isValidChargePath(adjacentTile) && adjacentTile.channels[channelOpposite] === true) { // if adjacentTile is valid/has channel connect
+          adjacentTile.chargeStatus.chargeAligned = true // change that tile to be charge aligned
+          inChargePath = adjacent[validChannels[i]] // set that tile as the new charge pathfinding start point and re-loop
+        }
+      }
+    }
+    if (lastChargedTile === inChargePath) {
+      inChargePath = null
+    }
+  }
+}
+
+function isValidChargePath(tile) {
+  return tile.chargeStatus.spent === false && tile.chargeStatus.chargeAligned === false && tile.chargeStatus.charged === false && (tile.isHidden === false || tile.sink === true)
+}
+
+module.exports = findChargePath
